refactor: split large files, add Sendable/MainActor, fix remaining issues
**File splits:**
- ListItem (695→305 lines) split into LemonadeSelectListItem,
  LemonadeResourceListItem, LemonadeActionListItem
- ToastManager (533→225 lines) split into LemonadeToastContainerView,
  LemonadeToastModifiers

**ToastManager fix:**
- Fix fire-and-forget Task race in exitCurrentToast — now stores and
  cancels previous exit task before creating new one

**Swift 6 readiness:**
- Add Sendable conformance to all value types: enums (LemonadeSpacing,
  LemonadeRadius, LemonadeSizes, LemonadeShadow, LemonadeToastVoice,
  LemonadeToastDuration, ToastAnimationConfig), structs (LemonadeTextStyle,
  LemonadeShadowData, LemonadeToastItem, impl structs), and protocols
  (BaseOpacity, StateOpacity, BaseBorderWidth, StateBorderWidth)
- Add Equatable to LemonadeToastDuration
- Add @MainActor to LemonadeDatePickerState and LemonadeDateRangePickerState

**SegmentedControl:**
- Make LemonadeTabButtonProperties conform to Identifiable + Hashable
- Change ForEach from id: \.offset to id: \.element.id for stable identity

**Other fixes:**
- Add missing border40 to CGFloat+Lemonade shorthand
- Add doc comment on LemonadeShadowData.spread explaining SwiftUI limitation
- Add test target to Package.swift with basic token value tests
